requirements.txt
click==8.0.1
Flask==2.0.1
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.1
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
PyJWT==1.7.1
SQLAlchemy==1.3.0
Werkzeug==2.0.1
psycopg2